https://www.hackerrank.com/challenges/vector-erase/problem?h_r=profile
You are provided with a vector of N integers. Then, you are given  2queries. For the first query, you are provided with  1 integer, which denotes a position in the vector. The value at this position in the vector needs to be erased. The next query consists of 2 integers denoting a range of the positions in the vector. The elements which fall under that range should be removed. The second query is performed on the updated vector which we get after performing the first query.